Cleaning and moving feature
Cleaning some of the code and moving the Disable Tracking to the ads section
