+++ app/src/main/java/ps/reso/instaeclipse/utils/feature/FeatureFlags.java
@@ -1,10 +1,13 @@
 package ps.reso.instaeclipse.utils.feature;
 
 public class FeatureFlags {
+
+    // Dev Options
     public static boolean isDevEnabled = false;
     public static boolean isImportingConfig = false;
     public static boolean isExportingConfig = false;
 
+    // Ghost Mode
     public static boolean isGhostModeEnabled = false;
     public static boolean isGhostSeen = false;
     public static boolean isGhostTyping = false;
@@ -22,6 +25,7 @@ public class FeatureFlags {
     public static boolean quickToggleLive = false;
 
 
+    // Distraction Free
     public static boolean isDistractionFree = false;
     public static boolean disableStories = false;
     public static boolean disableFeed = false;
@@ -29,15 +33,17 @@ public class FeatureFlags {
     public static boolean disableExplore = false;
     public static boolean disableComments = false;
 
+    // Ads and Analytics
     public static boolean isAdBlockEnabled = false;
     public static boolean isAnalyticsBlocked = false;
+    public static boolean disableTrackingLinks = false;
 
+    // Misc Options
     public static boolean isMiscEnabled = false;
     public static boolean disableStoryFlipping = false;
     public static boolean disableVideoAutoPlay = false;
     public static boolean showFollowerToast = false;
     public static boolean showFeatureToasts = false;
 
-    public static boolean disableTrackingLinks = false;
 
 }
